Ensure timeouts are called with the proper context.
Otherwise, you're invoking the public timeout methods on the exports
object, e.g.:

```javascript
require('timers').setTimeout(function() {
	// do things
}, 500);
```

will result in an Illegal invocation type error in chrome, because
`this` of `setTimeout` will point to the timers' module.exports
instead of window.

These are just pass-through methods that do no type checking, etc,
and apply the arguments directly to method with the context of
`window`.